27# More than one of the same upgrade -BUG-

+++ Assets/_Scripts/Interface_Attack.cs
@@ -3,6 +3,6 @@ using System.Collections.Generic;
 using UnityEngine;
 
 public interface Interface_Attack {
-	int getDamage();
+	float getDamage();
 	bool isActive();
 }
